Rendering works now on multiple screen resolution (not sure) Add first sound system iteration (using jogamp openAl)

+++ src/com/exiro/render/EntityRender.java
@@ -280,7 +280,7 @@ public class EntityRender {
         if (cs.getObject() != null) {
             cs.getObject().delete();
         }
-        for (int i = (int) Math.min(startX, currCase.getxPos()); i < Math.max(startX, currCase.getxPos()); i += defaultObject.getSize()) {
+        for (int i = Math.min(startX, currCase.getxPos()); i < Math.max(startX, currCase.getxPos()); i += defaultObject.getSize()) {
             for (int j = Math.min(startY, currCase.getyPos()); j < Math.max(startY, currCase.getyPos()); j += defaultObject.getSize()) {
                 Case c = GameManager.currentCity.getMap().getCase(i, j);
                 if (c != null && c.getObject() != null) {
